+++ RFID Timing/Device/Interfaces/DevControl.cs
@@ -3,246 +3,11 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using System.Net;
 using System.Runtime.InteropServices;
-using UHF;
-using System.Threading;
+
 
 namespace RFID_Timing
 {
-    class DeviceManager
-    {
-        public uint ip = 0;
-
-        public event EventHandler<TagCatchEventArgs> tagCatch;
-
-        public bool connectDevice()
-        {
-            //TODO: move this to thread
-            bool isConnected = false;
-            bool isFound = search();
-            if(isFound)
-            {
-                isConnected = connect();
-            }
-
-            return isConnected;
-        }
-
-        private bool search()
-        {
-            DevControl.tagErrorCode initErrorCode = DevControl.DM_Init(new SearchCallBack(searchHandler), IntPtr.Zero);
-
-            uint broadcastIp = 0xffffffff;
-            int timeout = 1500;
-            DevControl.tagErrorCode searchErrorCode = DevControl.DM_SearchDevice(broadcastIp, timeout);
-
-            if (searchErrorCode == DevControl.tagErrorCode.DM_ERR_OK)
-            {
-                return true;
-            }
-            else
-            {
-                return false;
-            }
-        }
-
-        private void searchHandler(IntPtr dev, IntPtr data)
-        {
-            StringBuilder devname = new StringBuilder(100);
-            StringBuilder macAdd = new StringBuilder(100);
-            DevControl.tagErrorCode eCode = DevControl.DM_GetDeviceInfo(dev, ref this.ip, macAdd, devname);
-        }
-
-        private bool connect()
-        {
-            string ipAddress = IPAddress.Parse(this.ip.ToString()).ToString();
-            int nPort = 27011;
-            byte fComAdr = 255;
-            int FrmPortIndex = 0;
-            int fCmdRet = RWDev.OpenNetPort(nPort, ipAddress, ref fComAdr, ref FrmPortIndex);
-            if (fCmdRet == 0)
-            {
-                RWDev.InitRFIDCallBack(new RFIDCallBack(RFIDTagCallback), true, FrmPortIndex);
-                return true;
-            }
-            else
-            {
-                Console.Write(getErrorCodeDesc(fCmdRet));
-                return false;
-            }
-        }
-
-        protected virtual void RFIDTagCallback(IntPtr p, Int32 nEvt)
-        {
-            RFIDTag rfidTag = (RFIDTag)Marshal.PtrToStructure(p, typeof(RFIDTag));
-            EventHandler<TagCatchEventArgs> handler = tagCatch;
-
-            TagCatchEventArgs e = new TagCatchEventArgs(){ Tag = rfidTag };
-            handler(this, e);
-        }
-
-        public void start()
-        {
-            Thread readerWorker = new Thread(listenReader);
-            readerWorker.IsBackground = true;
-            readerWorker.Start();
-        }
-
-        private void listenReader()
-        {
-            while (true)
-            {
-                byte fComAdr = 0;
-                //Multiply query parameters
-                byte Qvalue = 4;
-                byte Session = 255;
-
-                byte MaskMem = 0;
-                byte[] MaskAdr = new byte[2];
-                byte MaskLen = 0;
-                byte[] MaskData = new byte[100];
-                byte MaskFlag = 0;
-                byte AdrTID = 0;
-                byte LenTID = 6;
-                byte TIDFlag = 0;
-                byte Target = 0;
-                byte InAnt = 0x80;
-                byte Scantime = 20;
-                byte FastFlag = 0;
-                byte[] EPC = new byte[50000];
-                byte Ant = 0;
-                int Totallen = 0;
-                int TagNum = 0;
-
-                int frmcomportindex = 1;
-
-                int fCmdRet = RWDev.Inventory_G2(
-                    ref fComAdr,
-                    Qvalue,
-                    Session,
-                    MaskMem,
-                    MaskAdr,
-                    MaskLen,
-                    MaskData,
-                    MaskFlag,
-                    AdrTID,
-                    LenTID,
-                    TIDFlag,
-                    Target,
-                    InAnt,
-                    Scantime,
-                    FastFlag,
-                    EPC,
-                    ref Ant,
-                    ref Totallen,
-                    ref TagNum,
-                    frmcomportindex
-                );
-
-                Thread.Sleep(5);
-            }
-        }
-
-        private string getErrorCodeDesc(int code)
-        {
-            switch (code)
-            {
-                case 0x00:
-                case 0x26:
-                    return "success";
-                case 0x01:
-                    return "Return before Inventory finished";
-                case 0x02:
-                    return "the Inventory-scan-time overflow";
-                case 0x03:
-                    return "More Data";
-                case 0x04:
-                    return "Reader module MCU is Full";
-                case 0x05:
-                    return "Access Password Error";
-                case 0x09:
-                    return "Destroy Password Error";
-                case 0x0a:
-                    return "Destroy Password Error Cannot be Zero";
-                case 0x0b:
-                    return "Tag Not Support the command";
-                case 0x0c:
-                    return "Use the commmand,Access Password Cannot be Zero";
-                case 0x0d:
-                    return "Tag is protected,cannot set it again";
-                case 0x0e:
-                    return "Tag is unprotected,no need to reset it";
-                case 0x10:
-                    return "There is some locked bytes,write fail";
-                case 0x11:
-                    return "can not lock it";
-                case 0x12:
-                    return "is locked,cannot lock it again";
-                case 0x13:
-                    return "Parameter Save Fail,Can Use Before Power";
-                case 0x14:
-                    return "Cannot adjust";
-                case 0x15:
-                    return "Return before Inventory finished";
-                case 0x16:
-                    return "Inventory-Scan-Time overflow";
-                case 0x17:
-                    return "More Data";
-                case 0x18:
-                    return "Reader module MCU is full";
-                case 0x19:
-                    return "'Not Support Command Or AccessPassword Cannot be Zero";
-                case 0x1A:
-                    return "Tag custom function error";
-                case 0xF8:
-                    return "Check antenna error";
-                case 0xF9:
-                    return "Command execute error";
-                case 0xFA:
-                    return "Get Tag,Poor Communication,Inoperable";
-                case 0xFB:
-                    return "No Tag Operable";
-                case 0xFC:
-                    return "Tag Return ErrorCode";
-                case 0xFD:
-                    return "Command length wrong";
-                case 0xFE:
-                    return "Illegal command";
-                case 0xFF:
-                    return "Parameter Error";
-                case 0x30:
-                    return "Communication error";
-                case 0x31:
-                    return "CRC checksummat error";
-                case 0x32:
-                    return "Return data length error";
-                case 0x33:
-                    return "Communication busy";
-                case 0x34:
-                    return "Busy,command is being executed";
-                case 0x35:
-                    return "ComPort Opened";
-                case 0x36:
-                    return "ComPort Closed";
-                case 0x37:
-                    return "Invalid Handle";
-                case 0x38:
-                    return "Invalid Port";
-                case 0xEE:
-                    return "Return Command Error";
-                default:
-                    return "";
-            }
-        }
-
-    }
-
-    public class TagCatchEventArgs : EventArgs
-    {
-        public RFIDTag Tag { get; set; }
-    }
-
     internal delegate void SearchCallBack(IntPtr dev, IntPtr data);
 
     public class DevControl
@@ -689,4 +454,3 @@ namespace RFID_Timing
         }
     }
 }
-
