+++ RFID Timing/Events/TagCatch.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using UHF;
+
+namespace RFID_Timing
+{
+    public class TagCatchEventArgs : EventArgs
+    {
+        public RFIDTag Tag { get; set; }
+    }
+}
